Project

+++ src/main/java/com/Informatorio/InfoPrimeraApp/service/cancion/CancionServiceImpl.java
@@ -2,12 +2,18 @@ package com.Informatorio.InfoPrimeraApp.service.cancion;
 
 
 import com.Informatorio.InfoPrimeraApp.dominio.Cancion;
+import com.Informatorio.InfoPrimeraApp.dto.CancionDto;
+import com.Informatorio.InfoPrimeraApp.exception.NotFoundException;
+import com.Informatorio.InfoPrimeraApp.mapper.CancionMapper;
 import com.Informatorio.InfoPrimeraApp.repository.CancionRepository;
 import com.Informatorio.InfoPrimeraApp.service.cancion.CancionService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
+import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
@@ -16,20 +22,67 @@ public class CancionServiceImpl implements CancionService {
     private final CancionRepository cancionRepository;
 
     @Override
-    public Cancion crearCancion(Cancion cancion) {
+    public Cancion crearCancion(CancionDto cancionDto) {
+        Cancion cancion = CancionMapper.toCancion(cancionDto);
         return cancionRepository.save(cancion);
     }
 
     @Override
-    public Cancion obtenerCancionPorId(UUID id) {
-        return cancionRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Canción no encontrada con id: " + id));
+    public CancionDto obtenerCancionPorId(UUID id) {
+        Cancion cancion = cancionRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Canción", "ID", id.toString()));
+        return CancionMapper.toCancionDto(cancion);
     }
+
 
     @Override
     public void eliminarCancion(UUID id) {
+        if (!cancionRepository.existsById(id)) {
+            throw new NotFoundException("Canción", "ID", id.toString());
+        }
         cancionRepository.deleteById(id);
     }
 
-    // Implementación de otros métodos necesarios
+    @Override
+    public Cancion actualizarCancion(UUID idCancion, CancionDto cancionDto) {
+        Cancion cancionExistente = cancionRepository.findById(idCancion)
+                .orElseThrow(() -> new NotFoundException("Canción", "ID", idCancion.toString()));
+
+        return cancionRepository.save(cancionExistente);
+    }
+
+    @Override
+    public List<CancionDto> obtenerCancionesPorFiltro(String nombre, String genero, String artista) {
+        List<Cancion> canciones;
+
+        if (nombre != null && !nombre.isEmpty()) {
+            canciones = cancionRepository.findByNombreContaining(nombre);
+        } else if (genero != null && !genero.isEmpty()) {
+            canciones = cancionRepository.findByGeneroNombre(genero);
+        } else if (artista != null && !artista.isEmpty()) {
+            canciones = cancionRepository.findByArtistaNombre(artista);
+        } else {
+            canciones = cancionRepository.findAll();
+        }
+
+        return canciones.stream()
+                .map(CancionMapper::toCancionDto)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<CancionDto> obtenerTodasLasCanciones() {
+        return null;
+    }
+
+    public List<CancionDto> obtenerCancionesAleatorias() {
+        List<Cancion> todasLasCanciones = cancionRepository.findAll();
+        Collections.shuffle(todasLasCanciones);
+        return todasLasCanciones.stream().map(CancionMapper::toCancionDto).collect(Collectors.toList());
+    }
+
+    @Override
+    public List<CancionDto> buscarCanciones(String titulo, String genero, String artista, String album) {
+        return null;
+    }
 }
